+++ cmd/server/flags.go
@@ -38,13 +38,13 @@ var flags = append([]cli.Flag{
 		Aliases: []string{"log-xorm-sql"}, // TODO: remove in v4.0.0
 		Usage:   "enable logging of sql commands",
 	},
-	&cli.IntFlag{
+	&cli.Int64Flag{
 		Sources: cli.EnvVars("WOODPECKER_DATABASE_MAX_CONNECTIONS"),
 		Name:    "db-max-open-connections",
 		Usage:   "max connections xorm is allowed create",
 		Value:   100,
 	},
-	&cli.IntFlag{
+	&cli.Int64Flag{
 		Sources: cli.EnvVars("WOODPECKER_DATABASE_IDLE_CONNECTIONS"),
 		Name:    "db-max-idle-connections",
 		Usage:   "amount of connections xorm will hold open",
@@ -166,13 +166,13 @@ var flags = append([]cli.Flag{
 		Usage:   "The default docker image to be used when cloning the repo",
 		Value:   constant.DefaultClonePlugin,
 	},
-	&cli.IntFlag{
+	&cli.Int64Flag{
 		Sources: cli.EnvVars("WOODPECKER_DEFAULT_PIPELINE_TIMEOUT"),
 		Name:    "default-pipeline-timeout",
 		Usage:   "The default time in minutes for a repo in minutes before a pipeline gets killed",
 		Value:   60,
 	},
-	&cli.IntFlag{
+	&cli.Int64Flag{
 		Sources: cli.EnvVars("WOODPECKER_MAX_PIPELINE_TIMEOUT"),
 		Name:    "max-pipeline-timeout",
 		Usage:   "The maximum time in minutes you can set in the repo settings before a pipeline gets killed",
